+++ src/AutoDayTrader.Core/Spread.cs
@@ -2,12 +2,12 @@ namespace AutoDayTrader.Core;
 
 public class Spread
 {
-    public double Pips { get; }
-    public Spread(double pips)
+    public decimal Pips { get; }
+    public Spread(decimal pips)
     {
         this.Pips = pips;
     }
-    public static Spread From(Symbol symbol, double bid, double ask)
+    public static Spread From(Symbol symbol, decimal bid, decimal ask)
     {
         return new Spread(Math.Round((ask - bid) / symbol.Pip, 2));
     }
